refactor(EntityLink): changed function getEntityLinkJSX to component EntityLink
